requirements.txt
bokeh
flask
flask-sqlalchemy
flask-migrate
python-dotenv
pandas
